refactor: remove checkSignCondition from content.js
The resign info check has a much better performance and do not require any storage reading. More, checking resign condition every single time may increase the chance(although quit little) of missing one. It is also better and easier for debugging when human inspection is needed for check-in process. The page will close in milliseconds if it is opened by the plugin and hard to catch any action or log from it.